Moving the template slot option from the template to the config.
If it's in the template, then if we want to render that slot elsewhere (like when hitting cancel after editing it), we must have it in the config. There's no way to know that the slot should be rendered in the template if we're not rendering the full template itself unless it's in the config instead.